Version 0.4
Fixed: Entity queries not working properly if they are changed after initial use.
Fixed: Hashes not calculated twice anymore when creating new archetypes.
Changed: EnitityQueries recalculated on-demand
Changed: Component types sorted by type index now
Added: Iter benchmark for ForEach with internal query

+++ include/gaia/ecs/component.h
@@ -196,14 +196,14 @@ namespace gaia {
 
 		[[nodiscard]] inline uint64_t CalculateMatcherHash(std::span<const ComponentMetaData*> types) noexcept {
 			uint64_t hash = types.empty() ? 0 : types[0]->matcherHash;
-			for (uint32_t i = 1U; i < types.size(); ++i)
+			for (uint32_t i = 1U; i < (uint32_t)types.size(); ++i)
 				hash = utils::combine_or(hash, types[i]->matcherHash);
 			return hash;
 		}
 
 		[[nodiscard]] inline uint64_t CalculateLookupHash(std::span<const ComponentMetaData*> types) noexcept {
 			uint64_t hash = types.empty() ? 0 : types[0]->lookupHash;
-			for (uint32_t i = 1U; i < types.size(); ++i)
+			for (uint32_t i = 1U; i < (uint32_t)types.size(); ++i)
 				hash = utils::hash_combine(hash, types[i]->lookupHash);
 			return hash;
 		}
